Replaced all explicit partial handlers with Routing
