+++ src/Images/Api/Routing/ConceptIllustrationWarningPageRouting.cs
@@ -1,0 +1,30 @@
+﻿using System.Collections.Generic;
+using Simplified.Ring1;
+using Starcounter;
+
+namespace Images
+{
+    public class ConceptIllustrationWarningPageRouting : IPageRouting
+    {
+        public List<RoutingPreset> GetRoutingPresets()
+        {
+            return new List<RoutingPreset>
+            {
+                new RoutingPreset
+                {
+                    Uri = "/images/partials/imagewarning/{?}",
+                    PageCreator = info =>
+                    {
+                        var illustrationId = info.Arguments[0];
+                        var illustration = DbHelper.FromID(DbHelper.Base64DecodeObjectID(illustrationId)) as Illustration;
+                        if (illustration == null) return new Page();
+
+                        var page = new ConceptIllustrationWarningPage();
+                        page.RefreshData(illustration);
+                        return page;
+                    }
+                }
+            };
+        }
+    }
+}
